license-service/requirements/base.txt
Django==5.2.9
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
uuid6==2025.0.1
